methods implemented

+++ lab6/backend/src/main/java/com/example/backend/dto/InputParams.java
@@ -17,7 +17,7 @@ public class InputParams {
     double y0;
     double x0;
     double xn;
-    int h;
+    double h;
     double e;
